added a Clipping Function and a GetSAFURIfromexpofilepath.js function
now that ffmpeg is executing correctly we need to figure out how to get the file paths working correctly.  When I use the recording, the file path is in expo file system but when I use the ffmpeg it doesnt seem to be able to access files from their paths on my device and I cant find the SAF URI to try it out.  I need to find an absolute path or SAF URI in order to test whether the native module ffmpeg will be able to access the file that way or not.